+++ src/main/java/roomescape/PageViewController.java
@@ -1,0 +1,17 @@
+package roomescape;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class PageViewController {
+    @GetMapping("/")
+    public String showHomePage() {
+        return "home";
+    }
+
+    @GetMapping("/reservation")
+    public String showReservationOPage() {
+        return "reservation";
+    }
+}
