+++ src/main/java/roomescape/ReservationController.java
@@ -11,32 +11,23 @@ import java.util.List;
 
 @Controller
 @RequiredArgsConstructor
+@RequestMapping("/reservations")
 public class ReservationController {
 
     private final ReservationService reservationService;
 
-    @GetMapping("/")
-    public String showHomePage() {
-        return "home";
-    }
-
-    @GetMapping("/reservation")
-    public String showReservationOPage() {
-        return "reservation";
-    }
-
-    @GetMapping("/reservations")
+    @GetMapping("")
     public ResponseEntity<List<Reservation>> showReservationList() {
         return ResponseEntity.ok(reservationService.getAllReservations());
     }
 
-    @PostMapping("/reservations")
+    @PostMapping("")
     public ResponseEntity<Reservation> addReservation(@RequestBody Reservation reservation) {
         Reservation newReservation = reservationService.addReservation(reservation);
         return ResponseEntity.created(URI.create("/reservations/" + newReservation.getId())).body(newReservation);
     }
 
-    @DeleteMapping("/reservations/{id}")
+    @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteReservations(@PathVariable("id") Long id) {
         reservationService.deleteReservation(id);
         return ResponseEntity.noContent().build();
